feat(youtube): switch downloader defaults to yt-dlp
Migrate default binary names, updater artifact id, and command executor
from youtube-dl to yt-dlp while keeping the youtube plugin module id.
