+++ plugins/youtube/src/com/dabi/habitv/plugin/youtube/YoutubeConf.java
@@ -11,12 +11,20 @@ public final class YoutubeConf {
 	public static final String NAME = "youtube";
 	public static final String NAME_MP3 = "youtube-mp3";
 	public static final String ENCODING = "UTF-8";
+	/**
+	 * Default video download flags for yt-dlp (youtube-dl compatible subset).
+	 * Placeholders {@link com.dabi.habitv.framework.FrameworkConf#DOWNLOAD_INPUT} and
+	 * {@link com.dabi.habitv.framework.FrameworkConf#DOWNLOAD_DESTINATION} are resolved at runtime.
+	 */
 	public static final String DUMP_CMD = " \"#VIDEO_URL#\" -o \"#FILE_DEST#\"  --write-sub --write-auto-sub --no-check-certificate";
+	/**
+	 * Default MP3 extraction flags for yt-dlp ({@code --extract-audio} / {@code --audio-format}).
+	 */
 	public static final String DUMP_CMD_MP3 = " \"#VIDEO_URL#\" -o \"#FILE_DEST#\"  --extract-audio --audio-format mp3 --no-check-certificate";
 
 	public static final long MAX_HUNG_TIME = 300000L;
-	public static final String DEFAULT_WINDOWS_EXE = "youtube-dl.exe";
-	public static final String DEFAULT_LINUX_BIN_PATH = "youtube-dl";
+	public static final String DEFAULT_WINDOWS_EXE = "yt-dlp.exe";
+	public static final String DEFAULT_LINUX_BIN_PATH = "yt-dlp";
 	public static final String BASE_URL = "https://www.youtube.com";
 
 	public static String resolveApiKey() {
