+++ scripts/static-repo/tool-sources.properties
@@ -4,10 +4,10 @@
 #
 # type=url means download a single file; type=zip means download and extract packagedExeName from inner path
 #
-youtube-dl.type=url
-youtube-dl.version=latest
-youtube-dl.url=https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp.exe
-youtube-dl.packagedExeName=youtube-dl.exe
+yt-dlp.type=url
+yt-dlp.version=latest
+yt-dlp.url=https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp.exe
+yt-dlp.packagedExeName=yt-dlp.exe
 
 ffmpeg.type=zip
 ffmpeg.version=latest
